examples/packet_ordering: fix return value of parse_portmask
Giving invalid or zero portmask as command line option to
packet_ordering application will have an unexpected response.
The reason behind this is that parse_portmask's return value is
stored in a variable called portmask.The data type of portmask
is unsigned int, hence returning -1 in case of zero or invalid
portmask causes an unexpected behaviour.
If we return 0 instead of -1 this issue can be resolved.
The program already contains the functionality to print
"invalid portmask" and program usage if portmask is zero.

Fixes: 850f3733f840 ("examples/packet_ordering: new sample app")

Signed-off-by: Sarosh Arif <sarosh.arif@emumba.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>