eal: replace pci-whitelist/pci-blacklist options
Replace pci-whitelist with pci-allowlist and pci-blacklist with
pci-blocklist. Allow the old options for now, but make sure help
and all tests will use the new options.

The short flags are more commonly used and changing them
would impact more documentation and applications.
The -b flag can be described as "blocklist" rather than "blacklist.
The -w flag can be thought of as "with" instead "whitelist".

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>