net/mlx5: move GTP mask define to correct location
Recent patch [1] added definition of mask MLX5_GTP_FLAGS_MASK, just
above function flow_dv_validate_item_gtp(), where it is used.

Patch was applied together with other patches which modified the same
file, so the mask was located further away from the function it is
used in.

This patch moves the mask definition to the proper location.

Signed-off-by: Dekel Peled <dekelp@mellanox.com>
Acked-by: Matan Azrad <matan@mellanox.com>

[1] Commit 563ac307a46b ("net/mlx5: support match on GTP flags")

Signed-off-by: 0-day Robot <robot@bytheb.org>