lib/vhost: restrict pointer aliasing for packed path
Restrict pointer aliasing to allow the compiler to vectorize
loops more aggressively.

With this patch, a 9.6% improvement is observed in throughput for the
packed virtio-net PVP case, and a 2.8% improvement in throughput for
the packed virtio-user PVP case. All performance data are measured
under 0.001% acceptable packet loss with 1 core on both vhost and
virtio side.

Signed-off-by: Joyce Kong <joyce.kong@arm.com>
Reviewed-by: Phil Yang <phil.yang@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>