vhost: fix host notifier configuration error flow
A vDPA driver can configure its device FD to be notified directly by
the guest memory mapping using `rte_vhost_host_notifier_ctrl` API.

The driver request is managed by the dpdk vhost management and is
forwarded to the QEMU, the vhost massage includes reply request in order
to be sure that the memory mapping was done correctly by the QEMU.

When QEMU finishes the configuration, it marks that its replay is valid
in the slave FD using VHOST_USER_REPLY_MASK flag.
The flag is set only in success and when the slave FD includes the reply
data.

The vhost library didn't validate the above flag before accessing to the
slave FD, it leaded to the thread to be blocked on recvmsg call forever
in case the QEMU has some problems in the notifier configuration.

Handle VHOST_USER_REPLY_MASK flag to validate that slave FD includes
a reply data.

Fixes: d90cf7d111ac ("vhost: support host notifier")
Cc: stable@dpdk.org

Signed-off-by: Matan Azrad <matan@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>