net/mlx5: do not select legacy MPW implicitly
The Legacy MPW (multi-packet write) should not be engaged implicitly.
We should exclude this function form a Tx burst routine selection process
unless it is requested specifically by setting the txq_mpw_en devarg.
Exclude this function from the selection process the same way it is done
for the Enhanced MPW in the mlx5_select_tx_function() routine.

Fixes: eb8121ab9dac ("net/mlx5: introduce Tx burst routine template")
Cc: stable@dpdk.org

Signed-off-by: Alexander Kozyrev <akozyrev@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>