eal: replace usage of blacklist/whitelist in enum
The terms blacklist and whitelist are often seen as reminders
of the divisions in society. Instead, use more exact terms for
handling of which devices are used in DPDK.

This patch renames the enum values in the EAL include files.
As a backward compatiable temporary migration tool, define
a replacement mapping for old values.

The old names relating to blacklist and whitelist are replaced
by blocklist and allowlist, but applications may be using the
older compatiablity macros. To help with conversion to new names
cause a message when the compatabilty names are used.

This will be upgraded to a warning in the future.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>