net/ixgbe/base: modify loop accounting for retries
The condition for comparing retry against max_retry was flawed in the
do-while loops.  For the case where retry was initialized to 0 and
max_retry was initialized to 1, we'd break out of the loop at the
condition when the intent is to retry the code at least once.
Otherwise, the loop is unnecessary.  The other places have a larger
max_retry so code would get run multiple times (if necessary), but not
to the intended extent.

Signed-off-by: Jeb Cramer <jeb.j.cramer@intel.com>
Signed-off-by: Guinan Sun <guinanx.sun@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>