net/ixgbe/base: create dedicated func to restart auto nego
Code to restart auto negotiation on PHY is already used in two places
and is needed in CORE repo of ixgben. This patch places it in dedicated
function for code clarity and makes it available outside the SHARED repo.

Testing Hints (required if no HSD):
Check if link goes up and traffic works.
Change speed on the uplink.

Signed-off-by: Jakub Chylkowski <jakubx.chylkowski@intel.com>
Signed-off-by: Guinan Sun <guinanx.sun@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>