event/dlb: add shared code version 10.7.9
The DLB shared code is auto generated by Intel, and is being committed
here so that it can be built in the DPDK environment. The shared code
should not be modified. The shared code must be present in order to
successfully build the DLB PMD.

Changes since v1 patch series
1) convert C99 comment to standard C
2) remove TODO and FIXME comments

Signed-off-by: McDaniel, Timothy <timothy.mcdaniel@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>