app/testpmd: update uninitialized member when set fwd mode
When set the forwarding mode of the pkt, it needs to update
the fwd_eng that it is the member of cur_fwd_config. Otherwise,
an error will occur when executing the set command and show
confg fwd.
the flows as follows:
   1. ./tespmd -w xxx -l xx -n 4 -- i
   2. set fwd mac
      testpmd> set fwd mac
      Set mac packet forwarding mode

   3. show config fwd
      testpmd> show config fwd
      io packet forwarding - ports=1 - cores=1 - streams=8 -...
      Logical Core 1 (socket 0) forwards packets on 8 streams:

When run the 1 ~ 3 step, it should show the packet forwarding is
mac.

Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>