bus/pci: fix mmap PCI resource
When mapping a PCI BAR containing an MSI-X table, some devices do not
need to actually map this BAR or only need to map part of them, which
may cause the mapping to fail. Now some checks are added and a non-NULL
initial value is set to the variable to avoid this situation.

Fixes: 2fd3567e5425 ("pci: use OS generic memory mapping functions")
Cc: talshn@mellanox.com

Signed-off-by: Alvin Zhang <alvinx.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>