EAL: Called remove() of drivers for vdev and pci buses
while using memif with app, the resources are not cleaned on exit,
So an error occurred on running it second time. The cause of this problem
is that remove() of memif driver is not called by rte_eal_cleanup() which
is counterpart of probe() called from rte_eal_init(). This is a case for
all other divers e.g pci, so to solve this problem I have added the
functionality of calling remove() function of all the driver attached to
devices on vdev and pci buses.

Bugzilla ID: 353
Signed-off-by: Muhammad Bilal <m.bilal@emumba.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>