net/hns3: support keeping CRC
CRC is the end of frame, which occupies 4 bytes. Keeping CRC is a feature
of MAC, which will not strip CRC field when receiving frames. The featrue
can be enabled using DEV_RX_OFFLOAD_KEEP_CRC offload by upper level
application. And the feature is only supported for hns3 PF PMD driver,
not supported for hns3 VF PMD driver

Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: Wei Hu (Xavier) <xavier.huwei@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>