eal: replace usage of blacklist/whitelist in enum
The terms blacklist and whitelist are often seen as reminders
of the divisions in society. Instead, use more exact terms for
handling of which devices are used in DPDK.

This patch renames the enum values in the EAL include files.
As a backward compatible temporary migration tool, define
a replacement mapping for old values.

The old names relating to blacklist and whitelist are replaced
by blocked and allow, but applications may be using the
older compatibility macros. To help with conversion to new names
cause a message when the compatibility names are used.

The suffix PCI_ is also dropped because these arguments
can be used with other bus types (dpaa, vmbus, etc).

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Luca Boccassi <bluca@debian.org>
Acked-by: Gaetan Rivet <grive@u256.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>