net/i40e: fix incorrect hash look up table
The hash look up table(LUT) will not be initializing when starting
testpmd with --disable-rss. So that some invalid queue indexes may still
in the LUT. When enable RSS by creating RSS rule, some packets will not
be into the valid queues.
This patch fixes this issue by initializing the LUT when creating an RSS
rule.

Fixes: feaae285b342 ("net/i40e: support hash configuration in RSS flow")
Cc: stable@dpdk.org

Signed-off-by: Shougang Wang <shougangx.wang@intel.com>
Tested-by: Zhang, XiX <xix.zhang@intel.com>
Signed-off-by: Shougang Wang <shougangx.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>