bus/fslmc: support bus close API
This patch add support for closing the bus objects which
were acquired In the bus probe.

Some devices need to be cleaned while in both primary and
secondary process and while some need to be cleaned only in
case of primary process.

The devices are closed as per the white list used while
creating the objects in a particular process.

Signed-off-by: Rohit Raj <rohit.raj@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>