doc: announce move of aes gmac algorithm to aead
This patch announces removal of RTE_CRYPTO_AUTH_AES_GMAC
from rte_crypto_auth_algorithm and addition of RTE_CRYPTO_AEAD_AES_GMAC
to rte_crypto_aead_algorithm.
AES-GMAC is variation of AES-GCM algorithm with the difference that
it does not perform encryption. As a matter of fact internally
there is no difference between GMAC and GCM except for the way how
data is passed.
Moving GMAC to AEAD can simplify way of implementing this alogrithm
for example in IPsec (RFC4543).

Signed-off-by: Arek Kusztal <arkadiuszx.kusztal@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>